Tinygo support for internal/sysfs
